c-g: Rearrange check-graphite to be a multi-module bin
Separate argument parsing, assertion parsing/checking, and graphite data
fetching/massaging into their on modules.
